Add a new fixture that isolates filesystem
Since this program needs to read and write files, some tests must be
isolated to avoid any type of modification in the source code, files and
the GIT repository of this project.

The new fixture isolated_filesystem can be used as a context manager
to change the current working directory to a temporary directory created
by pytest.