Support version string update on selected files
Normally, in a python project, the current version string is defined
under a version or __version__ variable somewhere in a module. The version
string can now be updated on one or more files by passing the path of the
files in the end of the release subcommand.

USAGE:

    brau release setup.py package/__init__.py another_file.py

To avoid passing the files on each release, put the file paths
under the braulio section to the files key in the configuration
file.

!feat:release