Fix release markup being inserted in the wrong place
The previous algorithm looked for the main title in the changelog text
to insert the new content below. The problem is that it only recognized
titles without overline adornment, leading to wrong rendered rst files.

Another problem with this approach is that the changelog can not have
paragraphs between the main title and the release sections.

Therefore, from now on, the application will search for the title of the
latest version and insert the new content over it.

This should make the application work even with changelog that does not
have a main title.

!fix:changelog