Move label_pattern validation to a callback
``label_pattern`` depends on ``label_position``, and I didn't put the
validation on a callback because at that time I didn't fully understand
how a Click eager option works.

Setting ``label_position`` as eager, make it available to ``label_pattern``
callback, which is required in order to validate a given value propertly.

!refactor:release