Abort when a lower version is passed to --bump
At this time, releasing a version lower than the current version is not
supported. For example, if the current version is 4.0.5, version of the
3.y.z series can not be released.

This type of scenario must be addressed in a future. But for now, we
only show a message to the user telling that this behavior is not
supported.

!fix:release